refactor: extract shared ResourcePage models and histories
- move ResourcePage model types into core/weave/resource_page_models.ts
- move pure ResourcePage history grouping helpers into core/weave/resource_page_history_groups.ts
- keep core/weave/weave.ts as the stable re-export surface
- wire runtime page generation to the shared model and history modules
- preserve runtime WeaveRuntimeError wrapping for history parse failures
- update the codebase overview for the new module boundaries
