+++ documentation/notes/wd.codebase-overview.md
@@ -14,7 +14,7 @@ created: 1773673181726
   request/result types shared by all callers
   `core/targeting.ts` owns portable target request semantics: validate/generate target normalization, version-target normalization, prepared weave target derivation for version plus shared phases, root/recursive selection, and requested-target coverage rules including the intentional exact-versus-recursive asymmetry
   shared designator normalization now treats `/` as a CLI-only root sentinel and `""` as the internal root designator path, including root-aware target selection and support-artifact path derivation
-  `core/weave` has started splitting focused planners out of the large façade module; mesh support ResourcePage catch-up planning now lives in `mesh_support_pages.ts`, while `weave.ts` keeps the public re-export surface for existing runtime, CLI, and test imports
+  `core/weave` has started splitting focused planners and shared models out of the large façade module; mesh support ResourcePage catch-up planning now lives in `mesh_support_pages.ts`, shared ResourcePage model types live in `resource_page_models.ts`, pure ResourcePage history grouping lives in `resource_page_history_groups.ts`, while `weave.ts` keeps the public re-export surface for existing runtime, CLI, and test imports
   first Knop, first payload, and first extracted-Knop weave planning now resolve MeshInventory current/latest/next progression from `_mesh/_meta` instead of mutable current pointers in `_mesh/_inventory`; `_mesh/_inventory` keeps stable artifact-history and historical-state membership facts while `_mesh/_meta` advances `sflo:latestHistoricalState`, `sflo:nextStateOrdinal`, and consumed next-state hints
   current carried slices: `mesh create` request validation/support-artifact rendering, `knop create` planning over an existing mesh inventory, the first narrow `integrate` planning slice for `05-alice-knop-created-woven` -> `06-alice-bio-integrated`, the first narrow `knop add-reference` planning slice for `07-alice-bio-integrated-woven` -> `08-alice-bio-referenced`, the first narrow `payload.update` planning slice for `09-alice-bio-referenced-woven` -> `10-alice-bio-updated`, `extract` planning for both Alice Bio `11-alice-bio-v2-woven` -> `12-bob-extracted` and Fantasy Rules sidecar `07-shacl-integrated-woven` -> `08-ontology-and-shacl-terms-extracted`, and carried `weave` planning slices through Alice Bio `13-bob-extracted-woven` plus Fantasy Rules sidecar `15-first-release-woven`
