+++ src/core/weave/weave.ts
@@ -48,6 +48,14 @@ import {
   type WeaveSupportHistoryPolicies,
 } from "./support_history_policy.ts";
 import type { VersionPlan } from "./version_plan.ts";
+import type {
+  IdentifierResourcePageModel,
+  ReferenceCatalogCurrentLinkModel,
+  ReferenceCatalogResourcePageModel,
+  RepositorySourceFloatingLocator,
+  ResourcePageModel,
+  SimpleResourcePageModel,
+} from "./resource_page_models.ts";
 
 export { WeaveInputError } from "./errors.ts";
 export { planMeshSupportResourcePages } from "./mesh_support_pages.ts";
@@ -67,6 +75,25 @@ export type {
   WeaveSupportHistoryPolicies,
 } from "./support_history_policy.ts";
 export type { PlanMeshSupportResourcePagesInput } from "./mesh_support_pages.ts";
+export type {
+  CustomIdentifierRegionResourcePageModel,
+  CustomIdentifierResourcePageModel,
+  IdentifierResourcePageModel,
+  KnopArtifactLinkModel,
+  KnopResourcePageModel,
+  ReferenceCatalogCurrentLinkModel,
+  ReferenceCatalogResourcePageModel,
+  RepositorySourceFloatingLocator,
+  ResourcePageChildIdentifierModel,
+  ResourcePageExtractionSourceModel,
+  ResourcePageHistoryGroupModel,
+  ResourcePageHistoryStateModel,
+  ResourcePageModel,
+  ResourcePageRawSourcePanelModel,
+  ResourcePageReferenceLinkModel,
+  ResourcePageReferenceTargetModel,
+  SimpleResourcePageModel,
+} from "./resource_page_models.ts";
 export type { VersionPlan } from "./version_plan.ts";
 
 const RDF_TYPE_IRI = "http://www.w3.org/1999/02/22-rdf-syntax-ns#type";
@@ -167,11 +194,6 @@ export interface PayloadWorkingArtifact {
   latestHistoricalStatePath?: string;
 }
 
-export interface RepositorySourceFloatingLocator {
-  repositoryUrl: string;
-  repositoryPathFromRoot: string;
-}
-
 export interface ReferenceCatalogWorkingArtifact {
   workingLocalRelativePath: string;
   currentReferenceCatalogTurtle: string;
@@ -219,124 +241,6 @@ export interface WeaveableKnopCandidate {
   referenceTargetSourcePayloadArtifact?: ReferenceTargetSourcePayloadArtifact;
   resourcePageDefinitionArtifact?: ResourcePageDefinitionWorkingArtifact;
 }
-
-export interface IdentifierResourcePageModel {
-  kind: "identifier";
-  path: string;
-  designatorPath: string;
-  workingLocalRelativePath?: string;
-  workingAccessUrl?: string;
-  repositorySourceFloatingLocator?: RepositorySourceFloatingLocator;
-  extractionSource?: ResourcePageExtractionSourceModel;
-  references?: readonly ResourcePageReferenceLinkModel[];
-  childIdentifiers?: readonly ResourcePageChildIdentifierModel[];
-  historyGroups?: readonly ResourcePageHistoryGroupModel[];
-  rawSourcePanels?: readonly ResourcePageRawSourcePanelModel[];
-}
-
-export interface SimpleResourcePageModel {
-  kind: "simple";
-  path: string;
-  description: string;
-  childIdentifiers?: readonly ResourcePageChildIdentifierModel[];
-  historyGroups?: readonly ResourcePageHistoryGroupModel[];
-  rawSourcePanels?: readonly ResourcePageRawSourcePanelModel[];
-}
-
-export interface ReferenceCatalogCurrentLinkModel {
-  fragment: string;
-  referenceRoleLabel: string;
-  referenceTargetPath: string;
-  referenceTargetStatePath?: string;
-}
-
-export interface ReferenceCatalogResourcePageModel {
-  kind: "referenceCatalog";
-  path: string;
-  catalogPath: string;
-  ownerDesignatorPath: string;
-  currentLinks: readonly ReferenceCatalogCurrentLinkModel[];
-  historyGroups?: readonly ResourcePageHistoryGroupModel[];
-  rawSourcePanels?: readonly ResourcePageRawSourcePanelModel[];
-}
-
-export interface KnopArtifactLinkModel {
-  label: string;
-  path: string;
-}
-
-export interface KnopResourcePageModel {
-  kind: "knop";
-  path: string;
-  designatorPath: string;
-  ownerTitle?: string;
-  governedArtifacts: readonly KnopArtifactLinkModel[];
-  supportingArtifacts: readonly KnopArtifactLinkModel[];
-  childIdentifiers?: readonly ResourcePageChildIdentifierModel[];
-}
-
-export interface CustomIdentifierRegionResourcePageModel {
-  key: string;
-  markdown: string;
-  sourcePath: string;
-}
-
-export interface CustomIdentifierResourcePageModel {
-  kind: "customIdentifier";
-  path: string;
-  designatorPath: string;
-  definitionPath: string;
-  stylesheetPaths: readonly string[];
-  regions: readonly CustomIdentifierRegionResourcePageModel[];
-}
-
-export interface ResourcePageRawSourcePanelModel {
-  label: string;
-  sourcePath: string;
-  contents?: string;
-  omittedByteLength?: number;
-}
-
-export interface ResourcePageExtractionSourceModel {
-  sourceArtifactPath: string;
-  requestedTargetStatePath?: string;
-  artifactResolutionModeIri?: string;
-}
-
-export interface ResourcePageChildIdentifierModel {
-  label: string;
-  path: string;
-  rdfTypes?: readonly string[];
-}
-
-export interface ResourcePageReferenceTargetModel {
-  href: string;
-  label: string;
-}
-
-export interface ResourcePageReferenceLinkModel {
-  roleLabel: string;
-  targets: readonly ResourcePageReferenceTargetModel[];
-}
-
-export interface ResourcePageHistoryStateModel {
-  path: string;
-  manifestationPath?: string;
-  locatedFilePath?: string;
-}
-
-export interface ResourcePageHistoryGroupModel {
-  label: string;
-  path: string;
-  states: readonly ResourcePageHistoryStateModel[];
-}
-
-export type ResourcePageModel =
-  | IdentifierResourcePageModel
-  | KnopResourcePageModel
-  | SimpleResourcePageModel
-  | ReferenceCatalogResourcePageModel
-  | CustomIdentifierResourcePageModel;
 
 export interface PlanWeaveInput {
   request: VersionRequest;
